@@ -12,18 +12,18 @@
   "author": "lannex.shin@gmail.com",
   "license": "ISC",
   "dependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-password-strength-bar": "^0.3.5"
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-password-strength-bar": "^0.4.0"
   },
   "devDependencies": {
     "babel-eslint": "10.1.0",
-    "eslint": "7.25.0",
-    "eslint-config-airbnb": "18.2.1",
-    "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "7.23.2",
+    "eslint": "8.17.0",
+    "eslint-config-airbnb": "19.0.4",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-jsx-a11y": "6.5.1",
+    "eslint-plugin-react": "7.30.0",
     "parcel-bundler": "1.12.5"
   },
   "eslintConfig": {
