@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 // components
 import Home from './Home';
@@ -8,4 +8,7 @@ if (module.hot) {
   module.hot.accept();
 }
 
-ReactDOM.render(<Home />, document.getElementById('root'));
+const container = document.getElementById('root');
+
+const root = createRoot(container);
+root.render(<Home />);
